+++ dpeeg/datasets/__init__.py
@@ -15,3 +15,4 @@ from .gigadb import OpenBMI_MI
 from .gin import HighGamma
 from .physionet import PhysioNet_MI
 from .ukds import MODMA_128_Resting
+from .zenodo import Ofner2017